* Move ICSharpCode.Decompiler reference to Microsoft.CodeAnalysis.CSharp.EditorFeatures * Add Microsoft.CodeAnalysis.Editor.IDecompiledSourceService * Add basic implementation of the IDecompiledSourceService for C#: CSharpDecompiledSourceService * Move decompilation feature from MetadataAsSourceFileService to CSharpDecompiledSourceService
